Created ability inheritance. Added data and scripts for the second and third ability.
