+++ Assets/Data/AbilityData/SecondAbilityData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0a20a16690ccebc48aa651d7cb31be10
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
