+++ Assets/Data/AbilityData/ThirdAbilityData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d7125ab90b771e448e1a61221526702
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
