+++ Assets/Prefabs/Player/UI/Ability.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 6499452154846217437}
   - component: {fileID: 5039997177730243952}
   - component: {fileID: 5784263573581775771}
+  - component: {fileID: 945347023690376348}
   m_Layer: 5
   m_Name: Ability
   m_TagString: Untagged
@@ -76,6 +77,20 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!114 &945347023690376348
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 109236841743718588}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8a495d97bad61d74db849e94d1571abb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _cooldownDisplay: {fileID: 5491077960309337893}
+  _abilityImage: {fileID: 5784263573581775771}
 --- !u!1 &1842198563083731462
 GameObject:
   m_ObjectHideFlags: 0
