+++ Assets/Scripts/UI/Abilities/Ability.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 239ffd3a12e486447b0be74858a02b80
+guid: a558b0bfb6a070a4fa56ab39dd8c0ca6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
