+++ Assets/Scripts/UI/Abilities/DefaultAbility.cs
@@ -1,44 +1,6 @@
-using System.Collections;
-using TMPro;
 using UnityEngine;
 
-public class DefaultAbility : MonoBehaviour
+public class DefaultAbility : Ability
 {
-    [SerializeField] private AbilityData _abilityData;
-    [SerializeField] private AbilityDisplay _abilityDisplay;
 
-    private bool _isCooldownActive;
-
-    private void Start()
-    {
-        _abilityDisplay.SetAbilityImage(_abilityData.AbilityIcon);
-    }
-
-    public void AbilityCast()
-    {
-        if (_isCooldownActive)
-        {
-            return;
-        }
-        StartCoroutine(AbilityCooldown());
-    }
-
-    private IEnumerator AbilityCooldown()
-    {
-        _isCooldownActive = true;
-        float cooldown = _abilityData.AbilityCooldown;
-
-        _abilityDisplay.UpdateAbilityImage(true);
-
-        while (cooldown > 0)
-        {          
-            _abilityDisplay.UpdateCooldownDisplay(cooldown);
-            yield return new WaitForSeconds(1);
-            cooldown--;
-        }
-
-        _abilityDisplay.UpdateAbilityImage(false);
-
-        _isCooldownActive = false;
-    }
 }
