+++ Assets/Scripts/UI/Abilities/SecondAbility.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class SecondAbility : Ability
+{
+
+}
