+++ Assets/Scripts/UI/Abilities/SecondAbility.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 239ffd3a12e486447b0be74858a02b80
+guid: 6119702138235f3409e3e5e086cc0d5a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
