+++ Assets/Scripts/UI/Abilities/ThirdAbility.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class ThirdAbillity : Ability
+{
+
+}
