+++ Assets/Scripts/UI/Abilities/ThirdAbility.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 239ffd3a12e486447b0be74858a02b80
+guid: d7a5b3225c08f044a896252c18372694
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
