+++ Assets/Scripts/UI/Misc/PlayerInputHandler.cs
@@ -7,6 +7,8 @@ public class PlayerInputHandler : MonoBehaviour
     [SerializeField] private Menu _menu;
 
     [SerializeField] private DefaultAbility _defaultAbility;
+    [SerializeField] private SecondAbility _secondAbility;
+    [SerializeField] private ThirdAbillity _thirdAbillity;
 
     public void OnMouseInput(InputAction.CallbackContext context)
     {
@@ -29,6 +31,22 @@ public class PlayerInputHandler : MonoBehaviour
         if (context.started)
         {
             _defaultAbility.AbilityCast();
+        }
+    }
+
+    public void OnSecondAbility(InputAction.CallbackContext context)
+    {
+        if (context.started)
+        {
+            _secondAbility.AbilityCast();
+        }
+    }
+
+    public void OnThirdAbility(InputAction.CallbackContext context)
+    {
+        if (context.started)
+        {
+            _thirdAbillity.AbilityCast();
         }
     }
 }
